Allow posting to specific yammer group.
Yammer doesn't allow users to ignore specific users who post to the
toplevel company feed.  While you don't see these in the "My Feed"
stream, these messages do show up in the daily digest emails sent to
anyone who subscribes.

By containing this to a specific group, users can decide to watch the
project-specific tags, the group, or the integration user.  If a user
is not interested in development at all, this prevents the user from
being required to see it every day.