Add a Bugzilla service hook
This hook posts comments to a Bugzilla instance
over the XML-RPC web service when commit messages
reference a bug.

Muliple commits are grouped together and commits
may reference multiple bugs.

The hook checks that the commit hasn't already been
mentioned on the bug to avoid repeating a commit when
it's pushed to someone else's repository.