Bugzilla: Add a central repository option and support closing bugs
When the central repository option is set, including "fixes" or
"closes" in a commit message will close that bug when the commit is
pushed to the central repository.

Commits that have been mentioned in a bug when pushed to a user's
repository are repeated when pushed to the central repository, but
only the first line of the message is included.