Support HTTP and HTTPS connections to unfuddle, defaulting to HTTPS
Existing service configurations do not represent their preference for
HTTP over HTTPS, so it is assumed that they will need to continue using
HTTPS.

This change changes the logic to only switch the protocol from HTTPS to
HTTP if it is requested by the user configuring the service instance.
This makes the change opt-in and therefore backwards compatible with
existing data, while still providing the opportunity to switch over to
plain HTTP for API usage, which is really the only way to use the API
for the smaller Unfuddle plans.