Update the Bamboo service to work with Bamboo 4.x
The updates services uses the proper REST resource now.  The change includes
changing the authentication scheme from login -> trigger -> logout to use basic
auth when posting to the build queue resource. This reduces the number of HTTP
requests from n+2 to n (where n is the number of build plans that need to be
triggered).

This addresses issue #292