enable irc pull request functionality
Prior to this, the code existed for sending pull requests to IRC;
however, the code path wasn't enabled because the `default_events` was
only set to the Service default of `:push`. This commit enables the
existing code path by explicitly calling out that Service::IRC can
handle `:pull_request` events.

This commit also updates the documentation for the irc service to
reflect this new option, along with the previously available, but
undocumented, branch regular expressions field.