If domain included a trailing slash Packagist would reject the notification
If `http://packagist.org/` was entered, the URL being generated
was `http://packagist.org//api/github` which would fail as it
is not a valid route.

Fixed by replacing the strip command with lstrip followed by
a regex to replace any trailing slashes or whitespace with
nothing.

Added tests to ensure it is doing what is expected.

Reordered fields.