send a non-encoded token in authorizaton header
This Github service is currently asking the user for a direct access token from Heroku
in order to authenticate with its API. Heroku's API documentation specifies that, in the
case of direct authorization tokens, you have to send the Authorization header in a format
like this:

Authorization: Bearer [token]

This implies that you MUST NOT encode the token string. It has to be send as it the github
user supplied it to this service.

So, this commit changed that.
https://devcenter.heroku.com/articles/oauth#direct-authorization