Respect AWS region in AWS CodeDeploy hook
This fixes a bug that ignored the AWS region when using the AWS
CodeDeploy hook.  Previously, the us-east-1 region was always used,
unless a custom region was provided via a very custom deployment
payload.

When hosting code in a different region, the service hook fails with no
helpful error messages.  For example, see:
https://forums.aws.amazon.com/thread.jspa?messageID=597193

This fixes the code to use the aws_region configuration value.
Everything still defaults to us-east-1 if not set, and the custom
codedeploy payload is still supported.