Support for refs filtering in OBS
Introduce a new configuration parameter "refs" which contains a list
of wildcard patterns for Git reference. Hook should only trigger build
if the reference being pushed satisfy at least one of this patterns.

This would be useful when user do not want to rebuild their packages
on every commit pushed, but only for specific branches or tags.