Twitter: Preserve commit message whitespace
When tweeting commit messages, preserve the original commit message
whitespace and tweet it as-is.

Previously, String#split(" ") was used to process words, which splits on
*any* whitespace and with runs of contiguous whitespace characters
ignored [1]. When the split array was turned back into a string with
Array#join(" "), original whitespace characters were lost and replaced
by a single space character.

Now, String#gsub and a regular expression is used to process words,
which will preserve original whitespace.

Fixes #959.

[1] http://ruby-doc.org/core-2.2.0/String.html#method-i-split