irc.rb: Properly handle getting repo owner in PRs
Previously, it was assumed payload["repository"]["owner"]["name"] was
constant for all repositories. It was noticed that this is not the case,
and that the pull request event sends payload["repository"]["owner"]["login"]
instead.

A new method has been added to retrieve the repository owner, and
the tests have been amended to be more strict towards matching what
the realname may and may not contain.

fixes #1096, ref #1070