Remove CircleCI from Service Hooks integration
There was a time when we first migrated to the Webhooks API,
long ago, that we still accepted callbacks from hooks setup here.

We have since stopped accepting these hooks.

This is because there is no way to properly sign them here.

So the hooks the only way to use our service is with a signed hook which we created.

This has lead to some confusion, by users who accidentally delete our hook,
and try to recreate the hook themselves using this service.

We'd like to avoid support requests for these old hooks.