Ensure default region is used when none specified
The prior behavior caused region to be set to `""` when no region was
configured. This ended up causing a AWS `ops_works_region` to be set to `""`.
And, this ended up causing the `ops_works_endpoint` to be set to
`opsworks..amazonaws.com` instead of something like `opsworks.us-
east-1.amazonaws.com`. So, we work around this by ensuring `nil` is passed in
when no region is specified.