Fix sending Jabber notifications
This was broken due to an error that was received on sending a roster
message without an id. This was an issue in the xmpp4r-simple gem, so
this removes that gem and implements the necessary commands directly
with xmpp4r.

It removes the need for the existing monkey patch and also prevents
having to add even more. The code is not that much more complex by
implementing it directly.

Also removes the old test since it was not really exercising anything
since all the useful behavior was mocked. There isn't really anything
like VCR for XMPP to record real interactions and since github-services
is deprecated anyway, we don't want to invest too much time.