Add ability to set public key signature
This commit adds support for fetching and assigning a public key-based
signature of the request body.

Higher in the call stack, we instantiate a public key that is capable of
signing a message and inject it into the `Service` base class along with
whether or not the request needs to be signed. If it does need to be
signed, we add the base64 encoded signature to the headers of the
request, similar to HMAC signing.

This, paired with metadata that's injected into the webhook body with
requests that need to be signed, helps ensure that the request is
authentic and not a malicious payload and can help protect against
replay attacks.